Move test cases to newly created and specific test class
