+++ javalin/src/test/java/io/javalin/TestResponse.kt
@@ -20,7 +20,6 @@ import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.io.File
 import java.util.*
-import javax.servlet.http.Cookie
 
 class TestResponse {
 
@@ -122,47 +121,6 @@ class TestResponse {
         assertThat(http.call(HttpMethod.GET, "/hello-abs").body).isEqualTo("Redirected")
     }
 
-    @Test
-    fun `setting a cookie works`() = TestUtil.test { app, http ->
-        app.get("/create-cookie") { ctx -> ctx.cookie("Test", "Tast") }
-        app.get("/get-cookie") { ctx -> ctx.result(ctx.cookie("Test")!!) }
-        assertThat(http.get("/create-cookie").headers.getFirst(Header.SET_COOKIE)).isEqualTo("Test=Tast; Path=/")
-        assertThat(http.getBody("/get-cookie")).isEqualTo("Tast")
-    }
-
-    @Test
-    fun `setting a Cookie object works`() = TestUtil.test { app, http ->
-        app.get("/create-cookie") { ctx -> ctx.cookie(Cookie("Hest", "Hast").apply { maxAge = 7 }) }
-        assertThat(http.get("/create-cookie").headers.getFirst(Header.SET_COOKIE)).contains("Hest=Hast")
-        assertThat(http.get("/create-cookie").headers.getFirst(Header.SET_COOKIE)).contains("Max-Age=7")
-    }
-
-    @Test
-    fun `can't set duplicate cookies when other cookies set`() = TestUtil.test { app, http ->
-        app.get("/create-cookies") {
-            it.cookie("Test-1", "1")
-            it.cookie("Test-2", "2")
-            it.cookie("Test-3", "3")
-            it.cookie("Test-3", "4")  // duplicate
-        }
-        val response = http.get("/create-cookies");
-        assertThat(response.headers[Header.SET_COOKIE]!!).contains("Test-1=1; Path=/")
-        assertThat(response.headers[Header.SET_COOKIE]!!).contains("Test-2=2; Path=/")
-        assertThat(response.headers[Header.SET_COOKIE]!!).contains("Test-3=4; Path=/")
-        assertThat(response.headers[Header.SET_COOKIE]!!.size).isEqualTo(3)
-    }
-
-    @Test
-    fun `can't set duplicate cookies when no other cookies set`() = TestUtil.test { app, http ->
-        app.get("/create-cookies") {
-            it.cookie("MyCookie", "A")
-            it.cookie("MyCookie", "B")  // duplicate
-        }
-        val response = http.get("/create-cookies");
-        assertThat(response.headers[Header.SET_COOKIE]!!).contains("MyCookie=B; Path=/")
-        assertThat(response.headers[Header.SET_COOKIE]!!.size).isEqualTo(1)
-    }
-
     // Fix for https://github.com/tipsy/javalin/issues/543
     @Test
     fun `reading the result string resets the stream`() = TestUtil.test { app, http ->
@@ -177,15 +135,15 @@ class TestResponse {
     }
 
     private fun getSeekableInput(repeats: Int = SeekableWriter.chunkSize) = ByteArrayInputStream(
-            setOf("a", "b", "c").joinToString("") { it.repeat(repeats) }.toByteArray(Charsets.UTF_8)
+        setOf("a", "b", "c").joinToString("") { it.repeat(repeats) }.toByteArray(Charsets.UTF_8)
     )
 
     @Test
     fun `seekable - range works`() = TestUtil.test { app, http ->
         app.get("/seekable") { ctx -> ctx.seekableStream(getSeekableInput(), "text/plain") }
         val response = Unirest.get(http.origin + "/seekable")
-                .headers(mapOf(Header.RANGE to "bytes=${SeekableWriter.chunkSize}-${SeekableWriter.chunkSize * 2 - 1}"))
-                .asString().body
+            .headers(mapOf(Header.RANGE to "bytes=${SeekableWriter.chunkSize}-${SeekableWriter.chunkSize * 2 - 1}"))
+            .asString().body
         assertThat(response).doesNotContain("a").contains("b").doesNotContain("c")
     }
 
@@ -200,8 +158,8 @@ class TestResponse {
     fun `seekable - overreaching range works`() = TestUtil.test { app, http ->
         app.get("/seekable-3") { ctx -> ctx.seekableStream(getSeekableInput(), "text/plain") }
         val response = Unirest.get(http.origin + "/seekable-3")
-                .headers(mapOf(Header.RANGE to "bytes=0-${SeekableWriter.chunkSize * 4}"))
-                .asString().body
+            .headers(mapOf(Header.RANGE to "bytes=0-${SeekableWriter.chunkSize * 4}"))
+            .asString().body
         assertThat(response.length).isEqualTo(SeekableWriter.chunkSize * 3)
     }
 
@@ -209,8 +167,8 @@ class TestResponse {
     fun `seekable - file smaller than chunksize works`() = TestUtil.test { app, http ->
         app.get("/seekable-4") { ctx -> ctx.seekableStream(getSeekableInput(repeats = 50), "text/plain") }
         val response = Unirest.get(http.origin + "/seekable-4")
-                .headers(mapOf(Header.RANGE to "bytes=0-${SeekableWriter.chunkSize}"))
-                .asString().body
+            .headers(mapOf(Header.RANGE to "bytes=0-${SeekableWriter.chunkSize}"))
+            .asString().body
         assertThat(response.length).isEqualTo(150)
     }
